rename TypeInterface -> t_intf

+++ src/hash.h
@@ -1,8 +1,9 @@
 #ifndef _hash_h
 #define _hash_h
 
+#include <stdint.h>
 #include <stddef.h>
 
-unsigned long jenkins_hash(const void *obj, const size_t size);
+uint32_t jenkins_hash(const void *obj, const size_t size);
 
 #endif // _hash_h
